Refactored the Search logic into Service class and helpers.
