+++ source/VizGurka/Pages/Shared/_AnalyticsLayout.cshtml
@@ -168,4 +168,3 @@ private IHtmlContent RenderTree(Dictionary<string, object> tree, string parentId
     <script src="~/js/featureFilter.js" defer></script>
     @await RenderSectionAsync("Scripts", required: false)
 </body>
-</html>
